% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_chromatic.R
\name{chromatic_scale}
\alias{chromatic_scale}
\title{Chromatic scale constructor}
\usage{
chromatic_scale(
  aesthetics,
  scale_name,
  palette,
  name = waiver(),
  breaks = waiver(),
  n.breaks = NULL,
  labels = waiver(),
  limits = NULL,
  rescaler = rescale,
  oob = oob_censor,
  expand = waiver(),
  na.value = "grey50",
  trans = "identity",
  guide = "chromatic",
  prototype = NULL,
  channel_limits = NULL,
  super = ScaleChromatic
)
}
\arguments{
\item{aesthetics}{The names of the aesthetics that this scale works with.}

\item{scale_name}{The name of the scale that should be used for error messages
associated with this scale.}

\item{palette}{A palette function that when called with a \code{colour_spec}
vector should return a vector of colours.}

\item{name}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}

\item{breaks}{One of \itemize{
\item \code{NULL} for no breaks.
\item \code{waiver()} for the default breaks computed by the
\href{scales::trans_new()}{transformation object}.
\item A \code{colour_spec} vector. For continuous channels, must be a numeric
channel. For discrete channels, a character channel. Channels can be padded
with \code{NA}s if the desired breaks are of unequal length.
\item A \code{function} that uses the limits as input and returns breaks. Note
that this is used for both continuous and discrete channels.
\item A named \code{list} with the names of channels with (1) a \code{character}
or \code{numeric} vector giving the breaks for that channel or (2) a function to
be applied to the limits of that channel or (3) \code{NULL} for no breaks in
that channel. Channels whose names are absent in the \code{list}'s names are
treated with the \code{waiver()} option above.
}}

\item{n.breaks}{An integer guiding the number of major breaks. The algorithm
may choose a slightly different number to ensure nice break labels. Will
only have an effect if \code{breaks = waiver()}. Use \code{NULL} to use the default
number of breaks given by the transformation.}

\item{labels}{One of \itemize{
\item \code{NULL} for no labels.
\item \code{waiver()} for the default labels. In case of continuous channels,
these are passed through the format function of the
\href{scales::trans_new()}{transformation object}.
\item A \code{colour_spec} vector with character vectors in the channels. The
channels can be padded with \code{NA}s to match the length of channels with the
most breaks.
\item A \code{function} that uses the breaks as input and returns labels. Note
that this is used for both continuous and discrete channels.
\item A named \code{list} with the names of channels with (1) a \code{character}
vector giving the labels for that channel or (2) a function to be applied to
the breaks of that channel or (3) \code{NULL} for no labels in that channel.
Channels whose names are absent in the \code{list}'s names are treated with the
\code{waiver()} option above.
}}

\item{limits}{One of \itemize{
\item \code{NULL} to use the default scale range.
\item A \code{colour_spec} vector. For continuous channels, must be a length 2
vector giving the minimum and maximum. For discrete channels, the relevant
channel should define possible values. For mixed usage, the continuous
limits can be padded with \code{NA}s.
\item A \code{function} that accepts the existing (automatic) limits and returns
new limits. Note that this is used for both continuous and discrete
channels.
\item A named \code{list} with names of channels with (1) a vector defining
the limits or (2) a function to be applied to the natural limits. Channels
whose names are absent in the \code{list}'s names are treated with the \code{NULL}
option above.
}}

\item{rescaler}{A function used to scale the input values to the
range [0, 1]. This is always \code{\link[scales:rescale]{scales::rescale()}}, except for
diverging and n colour gradients (i.e., \code{\link[ggplot2:scale_gradient]{scale_colour_gradient2()}},
\code{\link[ggplot2:scale_gradient]{scale_colour_gradientn()}}). The \code{rescaler} is ignored by position
scales, which always use \code{\link[scales:rescale]{scales::rescale()}}.}

\item{oob}{One of:
\itemize{
\item Function that handles limits outside of the scale limits
(out of bounds).
\item The default (\code{\link[scales:oob]{scales::censor()}}) replaces out of
bounds values with \code{NA}.
\item \code{\link[scales:oob]{scales::squish()}} for squishing out of bounds values into range.
\item \code{\link[scales:oob]{scales::squish_infinite()}} for squishing infinite values into range.
}}

\item{expand}{For position scales, a vector of range expansion constants used to add some
padding around the data to ensure that they are placed some distance
away from the axes. Use the convenience function \code{\link[ggplot2:expansion]{expansion()}}
to generate the values for the \code{expand} argument. The defaults are to
expand the scale by 5\% on each side for continuous variables, and by
0.6 units on each side for discrete variables.}

\item{na.value}{Missing values will be replaced with this value.}

\item{trans}{For continuous scales, the name of a transformation object
or the object itself. Built-in transformations include "asn", "atanh",
"boxcox", "date", "exp", "hms", "identity", "log", "log10", "log1p", "log2",
"logit", "modulus", "probability", "probit", "pseudo_log", "reciprocal",
"reverse", "sqrt" and "time".

A transformation object bundles together a transform, its inverse,
and methods for generating breaks and labels. Transformation objects
are defined in the scales package, and are called \verb{<name>_trans} (e.g.,
\code{\link[scales:boxcox_trans]{scales::boxcox_trans()}}). You can create your own
transformation with \code{\link[scales:trans_new]{scales::trans_new()}}.}

\item{guide}{A function used to create a guide or its name. See
\code{\link[ggplot2:guides]{guides()}} for more information.}

\item{prototype}{A \code{function} that serves as constructor for the specific
\code{colour_spec} class.}

\item{channel_limits}{One of: \itemize{
\item A \code{colour_spec} vector of length 2 containing \code{numeric} channels that
indicating the limits for each channel between 0-1.
\item A named \code{list} with channel names and length 1 or 2 \code{numeric} vectors
that indicate the limits for that channel between 0-1.
}}

\item{super}{The super class to use for the constructed scale}
}
\value{
A \code{ScaleChromatic} ggproto object.
}
\description{
Constructs a chromatic scale. A chromatic scale can take several values for
the same observations and places these as coordinates in a colour space.
}
\examples{
NULL
}
