% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_cmy.R
\name{cmy_palette}
\alias{cmy_palette}
\title{CMY palette}
\usage{
cmy_palette(x, min = 0, max = 1)
}
\arguments{
\item{x}{A \code{cmy_spec} vector}

\item{min, max}{A \code{numeric(1)} specifying the lower and upper limits.}
}
\value{
A \code{character} vector of the same length as \code{x} with hexadecimal
colour notation.
}
\description{
This CMY palette converts a \code{cmy_spec} vector to colours.
}
\examples{
# By default, the limits are c(0, 1)
cmy_palette(cmy_spec(c(0, 0.5, 1), c(0.5, 0, 1), c(1, 0.5, 0)))
}
