% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs_cmy.R
\name{cmy_spec}
\alias{cmy_spec}
\title{Cyan-Magenta-Yellow specification}
\usage{
cmy_spec(c = double(), m = double(), y = double())
}
\arguments{
\item{c, m, y}{Vectors with values for the cyan, magenta and yellow channels
respectively.}
}
\value{
A \code{cmy_spec}, \code{vctrs_rcrd} S3 class object.
}
\description{
Creates a vector that stores values for cyan, magenta and yellow channels in
the CMY colour space.
}
\examples{
cmy_spec(c("A", "B", "C"), 1:3, 10:8)
}
\seealso{
Other colour space vectors: 
\code{\link{cmyk_spec}()},
\code{\link{hsl_spec}()},
\code{\link{hsv_spec}()},
\code{\link{rgb_spec}()}
}
\concept{colour space vectors}
