% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_cmyk.R
\name{cmyk_palette}
\alias{cmyk_palette}
\title{CMYK palette}
\usage{
cmyk_palette(x, min = 0, max = 1)
}
\arguments{
\item{x}{A \code{cmyk_spec} vector}

\item{min, max}{A \code{numeric(1)} specifying the lower and upper limits.}
}
\value{
A \code{character} vector of the same length as \code{x} with hexadecimal
colour notation.
}
\description{
This CMYK palette converts a \code{cmyk_spec} vector to colours.
}
\examples{
# By default, the limits are c(0, 1)
cmyk_palette(cmyk_spec(c(0, 0.5, 1), c(0.5, 0, 1),
                       c(1, 0.5, 0), c(0.2, 0.4, 0.6)))
}
