% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs_cmyk.R
\name{cmyk_spec}
\alias{cmyk_spec}
\title{Cyan-Magenta-Yellow-Black specification}
\usage{
cmyk_spec(c = double(), m = double(), y = double(), k = double())
}
\arguments{
\item{c, m, y, k}{Vectors with values for the cyan, magenta, yellow and key
(black) channels respectively.}
}
\value{
A \code{cmyk_spec}, \code{vctrs_rcrd} S3 class object.
}
\description{
Creates a vector that stores values for cyan, magenta, yellow and black
channels in the CMYK colour space.
}
\examples{
cmyk_spec(c("A", "B", "C"), 1:3, 10:8, c("X","Y","Z"))
}
\seealso{
Other colour space vectors: 
\code{\link{cmy_spec}()},
\code{\link{hsl_spec}()},
\code{\link{hsv_spec}()},
\code{\link{rgb_spec}()}
}
\concept{colour space vectors}
