% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs_colour_spaces.R
\name{colour_spec}
\alias{colour_spec}
\alias{hcl_spec}
\alias{hsv_spec}
\alias{hsl_spec}
\alias{rgb_spec}
\alias{cmyk_spec}
\alias{cmy_spec}
\title{Colour space vectors}
\usage{
hcl_spec(h = double(), c = double(), l = double())

hsv_spec(h = double(), s = double(), v = double())

hsl_spec(h = double(), s = double(), l = double())

rgb_spec(r = double(), g = double(), b = double())

cmyk_spec(c = double(), m = double(), y = double(), k = double())

cmy_spec(c = double(), m = double(), y = double())
}
\arguments{
\item{h, c, l, r, g, b, m, y, k}{Colour channels. Abbreviations are colour space
specific and written in full in the \strong{Functions} section. Can be a mix of
\code{logical}, \code{integer}, \code{double}, \code{character} or \code{factor} vectors of equal
length or length 1.}
}
\value{
An \code{colour_spec}, \code{vctrs_rcrd} S3 vector.
}
\description{
These functions create a vectors that store values specifying different
channels in a colour space.
}
\details{
These constructors are convenient for creating vectors that can be
supplied to the \code{aes()} function in ggplot2. During plot building, a
chromatic colour or fill scale will be assigned to coordinate the training
and mapping of colour space vectors. Missing channels will get placeholder
values that will not affect scale training.
}
\section{Functions}{
\itemize{
\item \code{hcl_spec}: Hue, Chroma and Luminance colour space.

\item \code{hsv_spec}: Hue, Saturation and Value colour space.

\item \code{hsl_spec}: Hue, Saturation and Lightness colour space.

\item \code{rgb_spec}: Red, Green and Blue colour space.

\item \code{cmyk_spec}: Cyan, Magenta, Yellow and Key (black) colour space.

\item \code{cmy_spec}: Cyan, Magenta and Yellow colour space.
}}

\examples{
NULL
}
\seealso{
\link[grDevices:rgb]{rgb()}, \link[grDevices:hsv]{hsv()},
\link[grDevices:hcl]{hcl()} and \code{\link[farver:encode_colour]{farver::encode_colour()}}.
}
