% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_rect.R
\name{guide_colourrect}
\alias{guide_colourrect}
\alias{guide_colorrect}
\title{Chromatic colour rectangle guide}
\usage{
guide_colourrect(
  title = waiver(),
  title.position = NULL,
  title.theme = NULL,
  title.hjust = NULL,
  title.vjust = NULL,
  label = TRUE,
  label.position = NULL,
  label.theme = NULL,
  label.hjust = NULL,
  label.vjust = NULL,
  check.overlap = TRUE,
  rectwidth = NULL,
  rectheight = NULL,
  nbin = 50,
  raster = TRUE,
  frame.colour = "black",
  frame.linewidth = 0.5,
  frame.linetype = 1,
  ticks = TRUE,
  ticks.colour = "black",
  ticks.linewidth = 0.5,
  default.unit = "line",
  channels = "auto",
  order = 0,
  available_aes = c("colour", "color", "fill"),
  ...
)

guide_colorrect(
  title = waiver(),
  title.position = NULL,
  title.theme = NULL,
  title.hjust = NULL,
  title.vjust = NULL,
  label = TRUE,
  label.position = NULL,
  label.theme = NULL,
  label.hjust = NULL,
  label.vjust = NULL,
  check.overlap = TRUE,
  rectwidth = NULL,
  rectheight = NULL,
  nbin = 50,
  raster = TRUE,
  frame.colour = "black",
  frame.linewidth = 0.5,
  frame.linetype = 1,
  ticks = TRUE,
  ticks.colour = "black",
  ticks.linewidth = 0.5,
  default.unit = "line",
  channels = "auto",
  order = 0,
  available_aes = c("colour", "color", "fill"),
  ...
)
}
\arguments{
\item{title}{A character string or expression indicating the title of guide.
If \code{NULL}, the title is not shown. By default (\code{waiver()}), the name of the
scale object or the name specified in \code{labs()} is used for the title. Note
that the colour rectangle guide can take 2 titles: one for each axis.}

\item{title.position}{A character string indicating the position of a
title. One of "top" (default for a vertical guide), "bottom", "left"
(default for a horizontal guide), or "right."}

\item{title.theme}{A theme object for rendering the title text. Usually the
object of \code{\link[ggplot2:element]{element_text()}} is expected. By default, the theme is
specified by \code{legend.title} in \code{\link[ggplot2:theme]{theme()}} or theme.}

\item{title.hjust}{A number specifying horizontal justification of the
title text.}

\item{title.vjust}{A number specifying vertical justification of the title
text.}

\item{label}{logical. If \code{TRUE} then the labels are drawn. If
\code{FALSE} then the labels are invisible.}

\item{label.position}{A character string indicating the position of a
label. One of "top", "bottom" (default for horizontal guide), "left", or
"right" (default for vertical guide).}

\item{label.theme}{A theme object for rendering the label text. Usually the
object of \code{\link[ggplot2:element]{element_text()}} is expected. By default, the theme is
specified by \code{legend.text} in \code{\link[ggplot2:theme]{theme()}}.}

\item{label.hjust}{A numeric specifying horizontal justification of the
label text.}

\item{label.vjust}{A numeric specifying vertical justification of the label
text.}

\item{check.overlap}{If \code{TRUE}, overlapping labels are silently removed. If
\code{FALSE}, labels are displayed regardless of whether they overlap.}

\item{rectwidth, rectheight}{A \code{numeric(1)} or \code{grid::unit()} object
specifying the width/height of the colour rectangle. Default value is the
\code{legend.key.width/height} or \code{legend.key.size} in the theme, times 4.}

\item{nbin}{A numeric specifying the number of bins for drawing the
colourbar. A smoother colourbar results from a larger value.}

\item{raster}{A logical. If \code{TRUE} then the colourbar is rendered as a
raster object. If \code{FALSE} then the colourbar is rendered as a set of
rectangles. Note that not all graphics devices are capable of rendering
raster image.}

\item{frame.colour}{A string specifying the colour of the frame
drawn around the bar. If \code{NULL} (the default), no frame is drawn.}

\item{frame.linewidth}{A numeric specifying the width of the frame
drawn around the bar.}

\item{frame.linetype}{A numeric specifying the linetype of the frame
drawn around the bar.}

\item{ticks}{A logical specifying if tick marks on the colourbar should be
visible.}

\item{ticks.colour}{A string specifying the colour of the tick marks.}

\item{ticks.linewidth}{A numeric specifying the width of the tick marks.}

\item{default.unit}{A character string indicating \code{\link[grid:unit]{grid::unit()}}
for \code{barwidth} and \code{barheight}.}

\item{channels}{A \code{character()} with the channel names to display. The
default, \code{"auto"}, removes empty channels and subsequently takes the first
two.}

\item{order}{positive integer less than 99 that specifies the order of
this guide among multiple guides. This controls the order in which
multiple guides are displayed, not the contents of the guide itself.
If 0 (default), the order is determined by a secret algorithm.}

\item{available_aes}{A vector of character strings listing the aesthetics
for which a colourbar can be drawn.}

\item{...}{ignored.}
}
\value{
A \code{guide_colourrect} S3 object.
}
\description{
The colour rectangle guide is a specialised guide for chromatic scales. It
maps two channels of a chromatic scales along the x and y axes and renders a
rectangle raster displaying the colours.
}
\examples{
# Make basic plot
g <- ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = rgb_spec(as.factor(cyl), drat, qsec)))

# Choose the channels
g + guides(colour = guide_colourrect(channels = c("g", "b")))

# Titles can take length 2 for the x- and y-axis.
g + guides(
  colour = guide_colourrect(title = c("# Cylinders", "Rear axle ratio"))
)

# If only 2 channels are specified, `guide_colourrect()` becomes the guide
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = hcl_spec(drat, qsec)))
}
\seealso{
Other guides for chromatic scales: 
\code{\link{guide_chromatic}()},
\code{\link{guide_colourbar2}()},
\code{\link{guide_colourcube}()}
}
\concept{guides for chromatic scales}
