% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_hsl.R
\name{hsl_palette}
\alias{hsl_palette}
\title{HSL palette}
\usage{
hsl_palette(x, min = 0, max = 1)
}
\arguments{
\item{x}{A \code{hsl_spec} vector}

\item{min, max}{A \code{numeric(1)} specifying the upper and lower limits}
}
\value{
A \code{character} vector of the same length as \code{x} with hexadecimal
colour notation.
}
\description{
This HSL palette converts a \code{hsl_spec} vector to colours.
}
\examples{
# By default, the limits are c(0, 1)
hsl_palette(hsl_spec(c(0, 0.5, 1), c(0.5, 0, 1), c(1, 0.5, 0)))
}
