% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs_hsl.R
\name{hsl_spec}
\alias{hsl_spec}
\title{Hue-Saturation-Lightness specification}
\usage{
hsl_spec(h = double(), s = double(), l = double())
}
\arguments{
\item{h, s, l}{Vectors with values for hue, saturation and lightness channels
respectively}
}
\value{
A \code{hsl_spec}, \code{vctrs_rcrd} S3 class object.
}
\description{
Creates a vector that stores values for hue, saturation and lightness
channels in the HSL colour space.
}
\examples{
hsv_spec(c("A", "B", "C"), c(1, 2, 3), c(10, 9, 8))
}
\seealso{
Other colour space vectors: 
\code{\link{cmyk_spec}()},
\code{\link{hsv_spec}()},
\code{\link{rgb_spec}()}
}
\concept{colour space vectors}
