% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs_hsv.R
\name{hsv_spec}
\alias{hsv_spec}
\title{Hue-Saturation-Value specification}
\usage{
hsv_spec(h = double(), s = double(), v = double())
}
\arguments{
\item{h, s, v}{Vectors with values for the hue, saturation and value channels
respectively.}
}
\value{
A \code{hsv_spec}, \code{vctrs_rcrd} S3 class object.
}
\description{
Creates a vector that stores values for hue, saturation and value channels in
the HSV colour space.
}
\examples{
hsv_spec(c("A", "B", "C"), c(1, 2, 3), c(10, 9, 8))
}
