% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad_nas}
\alias{pad_nas}
\title{Pad with NAs}
\usage{
pad_nas(x, size)
}
\arguments{
\item{x}{The object to pad.}

\item{size}{To what length to pad.}
}
\value{
A vector of length \code{size} or list wherein elements are of length
\code{size}.
}
\description{
This is a simple helper for some of the internals.
}
\examples{
pad_nas(1, 2)
}
