% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_rgb.R
\name{rgb_palette}
\alias{rgb_palette}
\title{RGB palette}
\usage{
rgb_palette(x, min = 0, max = 1)
}
\arguments{
\item{x}{A \code{rgb_spec} vector}

\item{min, max}{A \code{numeric(1)} specifying the upper and lower limits.}
}
\value{
A \code{character} vector of the same length as \code{x} with hexadecimal
colour notation.
}
\description{
This RGB palette converts a \code{rgb_spec} vector to colours.
}
\examples{
# By default, the limits are c(0, 1)
rgb_palette(rgb_spec(c(0, 0.5, 1), c(0.5, 0, 1), c(1, 0.5, 0)))
}
