% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs_rgb.R
\name{rgb_spec}
\alias{rgb_spec}
\title{Red-Green-Blue specification}
\usage{
rgb_spec(r = double(), g = double(), b = double())
}
\arguments{
\item{r, g, b}{Vectors with values for the red, green and blue channels
respectively.}
}
\value{
A \code{rgb_spec}, \code{vctrs_rcrd} S3 class object.
}
\description{
Creates a vector that stores values for red, green and blue channels in the
RGB colour space.
}
\examples{
rgb_spec(c("A", "B", "C"), c(1, 2, 3), c(10, 9, 8))
}
