% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_rgb.R
\name{scale_rgb}
\alias{scale_rgb}
\alias{scale_colour_rgb}
\alias{scale_fill_rgb}
\title{Red green blue colour scale}
\usage{
scale_colour_rgb(
  ...,
  na.value = 0,
  guide = "none",
  limits = NULL,
  aesthetics = c("colour", "red", "green", "blue")
)

scale_fill_rgb(
  ...,
  na.value = 0,
  guide = "none",
  limits = NULL,
  aesthetics = c("fill", "red", "green", "blue")
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=chromatic_scale]{chromatic_scale}}
  \describe{
    \item{\code{palette}}{A palette function that when called with a \code{colour_spec}
vector should return a vector of colours.}
    \item{\code{breaks}}{NOT IMPLEMENTED YET}
    \item{\code{labels}}{NOT IMPLEMENTED YET}
    \item{\code{prototype}}{A \code{function} that serves as constructor for the specific
\code{colour_spec} class.}
    \item{\code{channel_limits}}{A \code{colour_spec} vector of length 2 indicating the
limits for each channel.}
    \item{\code{scale_name}}{The name of the scale that should be used for error messages
associated with this scale.}
    \item{\code{name}}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}
    \item{\code{n.breaks}}{An integer guiding the number of major breaks. The algorithm
may choose a slightly different number to ensure nice break labels. Will
only have an effect if \code{breaks = waiver()}. Use \code{NULL} to use the default
number of breaks given by the transformation.}
    \item{\code{rescaler}}{A function used to scale the input values to the
range [0, 1]. This is always \code{\link[scales:rescale]{scales::rescale()}}, except for
diverging and n colour gradients (i.e., \code{\link[ggplot2:scale_gradient]{scale_colour_gradient2()}},
\code{\link[ggplot2:scale_gradient]{scale_colour_gradientn()}}). The \code{rescaler} is ignored by position
scales, which always use \code{\link[scales:rescale]{scales::rescale()}}.}
    \item{\code{oob}}{One of:
\itemize{
\item Function that handles limits outside of the scale limits
(out of bounds).
\item The default (\code{\link[scales:oob]{scales::censor()}}) replaces out of
bounds values with \code{NA}.
\item \code{\link[scales:oob]{scales::squish()}} for squishing out of bounds values into range.
\item \code{\link[scales:oob]{scales::squish_infinite()}} for squishing infinite values into range.
}}
    \item{\code{expand}}{For position scales, a vector of range expansion constants used to add some
padding around the data to ensure that they are placed some distance
away from the axes. Use the convenience function \code{\link[ggplot2:expansion]{expansion()}}
to generate the values for the \code{expand} argument. The defaults are to
expand the scale by 5\% on each side for continuous variables, and by
0.6 units on each side for discrete variables.}
    \item{\code{trans}}{For continuous scales, the name of a transformation object
or the object itself. Built-in transformations include "asn", "atanh",
"boxcox", "date", "exp", "hms", "identity", "log", "log10", "log1p", "log2",
"logit", "modulus", "probability", "probit", "pseudo_log", "reciprocal",
"reverse", "sqrt" and "time".

A transformation object bundles together a transform, its inverse,
and methods for generating breaks and labels. Transformation objects
are defined in the scales package, and are called \verb{<name>_trans} (e.g.,
\code{\link[scales:boxcox_trans]{scales::boxcox_trans()}}). You can create your own
transformation with \code{\link[scales:trans_new]{scales::trans_new()}}.}
    \item{\code{super}}{The super class to use for the constructed scale}
  }}

\item{na.value}{Missing values will be replaced with this value.}

\item{guide}{A function used to create a guide or its name. See
\code{\link[ggplot2:guides]{guides()}} for more information.}

\item{limits}{One of \itemize{
\item \code{NULL} to use the default scale range
\item A \code{colour_spec} vector. For continuous channels, must be a length 2
vector giving the minimum and maximum. For discrete channels, the relevant
channel should define possible values. For mixed usage, the continuous
limits can be padded with \code{NA}s.
\item A \code{function} that accepts the existing (automatic) limits and returns
new limits.
\item A named \code{list} with names of the channels with one of the above per
channel.
}}

\item{aesthetics}{The names of the aesthetics that this scale works with.}
}
\value{
A \code{ScaleChromatic} ggproto object that can be added to a plot.
}
\description{
Maps the values of several variables to coordinates in the RGB colour space.
}
\examples{
NULL
}
