% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{without_nas}
\alias{without_nas}
\title{Remove the NAs}
\usage{
without_nas(x)
}
\arguments{
\item{x}{A vector or \code{list} to remove the \code{NA}s from.}
}
\value{
A vector or \code{list} without \code{NA}s.
}
\description{
This is a simple helper for some of the internals.
}
\examples{
without_nas(c(NA, 2))
}
\keyword{internal}
