% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs_colour_spaces.R
\name{colour_spec}
\alias{colour_spec}
\alias{hcl_spec}
\alias{hsv_spec}
\alias{hsl_spec}
\alias{rgb_spec}
\alias{cmyk_spec}
\alias{cmy_spec}
\alias{lab_spec}
\alias{oklab_spec}
\title{Colour space vectors}
\usage{
hcl_spec(h = double(), c = double(), l = double())

hsv_spec(h = double(), s = double(), v = double())

hsl_spec(h = double(), s = double(), l = double())

rgb_spec(r = double(), g = double(), b = double())

cmyk_spec(c = double(), m = double(), y = double(), k = double())

cmy_spec(c = double(), m = double(), y = double())

lab_spec(l = double(), a = double(), b = double())

oklab_spec(l = double(), a = double(), b = double())
}
\arguments{
\item{h, c, l, s, v, r, g, b, m, y, k, a}{Colour channels. Abbreviations are colour
space specific and written in full in the \strong{Functions} section. Can be a
mix of \code{logical}, \code{integer}, \code{double}, \code{character} or \code{factor} vectors of
equal length or length 1.}
}
\value{
An \code{colour_spec}, \code{vctrs_rcrd} S3 vector.
}
\description{
These functions create vectors that store values specifying different
channels in a colour space.
}
\details{
These constructors are convenient for creating vectors that can be
supplied to the \code{aes()} function in ggplot2. During plot building, a
chromatic colour or fill scale will be assigned to coordinate the training
and mapping of colour space vectors. Missing channels will get placeholder
values that will not affect scale training.

The 'rule' is that your can combine two vectors of the same colour space,
but only if the colour channels can be combined. For example, you can
combine \code{rgb_spec(10, 5, "A")} with \code{rgb_spec(6, 2, "B")}. You can not
combine \code{hcl_spec(7, "C", 12)} with neither \code{hcl_spec(12, 8, 9)} nor a
different colour space \code{cmy_spec(3, "D", 4)}.
}
\section{Functions}{
\itemize{
\item \code{hcl_spec}: Hue, Chroma and Luminance colour space.

\item \code{hsv_spec}: Hue, Saturation and Value colour space.

\item \code{hsl_spec}: Hue, Saturation and Lightness colour space.

\item \code{rgb_spec}: Red, Green and Blue colour space.

\item \code{cmyk_spec}: Cyan, Magenta, Yellow and Key (black) colour space.

\item \code{cmy_spec}: Cyan, Magenta and Yellow colour space.

\item \code{lab_spec}: Lightness*, a* (green-red axis), b* (blue-yellow
axis) colour space. Also known as CIE L*ab.

\item \code{oklab_spec}: An OK Lightness, a (green-red axis), b (blue-yellow
axis) colour space. Also known as OKLab. Needs farver package version
\>2.0.3.
}}

\examples{
# In combination with ggplot2
ggplot(mtcars, aes(mpg, disp)) +
 geom_point(aes(colour = hsv_spec(mpg, drat, wt)))

# Concatenation
c(rgb_spec(1:2, c("A", "B"), 3:2), rgb_spec(12:11, c("", "X"), 4:5))

# Convert all channels to numeric matrix with `as.matrix()`
as.matrix(hcl_spec(10:15, LETTERS[1:6], 12:7))

# Using colour_spec vectors with palettes
cmyk_palette(cmyk_spec(0.1, 0.2, 0.3, 0))
}
\seealso{
\link{scale_chromatic} for the colour and fill scales that mirror these
colour space vectors.

\link[grDevices:rgb]{rgb()}, \link[grDevices:hsv]{hsv()},
\link[grDevices:hcl]{hcl()} and \code{\link[farver:encode_colour]{farver::encode_colour()}}.
}
