% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggchromatic-package.R
\docType{package}
\name{ggchromatic-package}
\alias{ggchromatic}
\alias{ggchromatic-package}
\title{ggchromatic: Colourspace Scales for ggplot2}
\description{
The 'ggchromatic' package provides additional colour and fill 
  scales to use with 'ggplot2'. It uses the 'ggplot2' extension system to map 
  a number of variables to different colourspaces with the 'farver' package.
}
\author{
\strong{Maintainer}: Teun van den Brand \email{tahvdbrand@gmail.com} (\href{https://orcid.org/0000-0002-9335-7468}{ORCID})

}
\keyword{internal}
