% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_chromatic.R
\name{guide_chromatic}
\alias{guide_chromatic}
\title{Chromatic guide}
\usage{
guide_chromatic(
  title = waiver(),
  ...,
  available_aes = c("colour", "color", "fill")
)
}
\arguments{
\item{title}{A character string or expression indicating a title of guide.
If \code{NULL}, the title is not shown. By default
(\code{\link[ggplot2:waiver]{waiver()}}), the name of the scale object or the name
specified in \code{\link[ggplot2:labs]{labs()}} is used for the title.}

\item{...}{Captures arguments to pass down to the constructor of the chosen
guide.}

\item{available_aes}{A vector of character strings listing the aesthetics
for which a colourbar can be drawn.}
}
\value{
A \code{chromatic_guide} S3 object.
}
\description{
This 'guide' is not really a guide, but a decision function for chromatic
scales. It chooses the guide based on the number of channels a chromatic
scale has seen data from. It is the default guide for the chromatic scales.
}
\details{
This 'guide' chooses in the following ways based on the number of
channels for which data is available.
\describe{
\item{0 channels}{chooses \code{guide_none()}}
\item{1 channel}{chooses \code{guide_colourbar2()}}
\item{2 channels}{chooses \code{guide_colourrect()}}
\item{3 channels}{chooses \code{guide_colourcube()}}
\item{More channels}{chooses \code{guide_colourcube()} with 3 channels}
}
}
\examples{
# Setup example plot
df <- data.frame(
  x = c(row(volcano)), y = c(col(volcano)), z = c(volcano)
)
g <- ggplot(df, aes(x, y)) +
  guides(fill = guide_chromatic())

# When the colour space has 3 defined channels, it makes a cube
g + geom_raster(aes(fill = cmy_spec(x, y, z)))

# When 2 channels are defined, it makes a rectangle
g + geom_raster(aes(fill = cmy_spec(x, z)))

# One defined channel gives a colour bar
g + geom_raster(aes(fill = cmy_spec(z)))
}
\seealso{
The \link{scale_chromatic} page for chromatic scales.

Other guides for chromatic scales: 
\code{\link{guide_colourbar2}()},
\code{\link{guide_colourcube}()},
\code{\link{guide_colourrect}()}
}
\concept{guides for chromatic scales}
