% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_cube.R
\name{guide_colourcube}
\alias{guide_colourcube}
\title{A character string or expression indicating the title of guide. If
\code{NULL}, the title is not shown. By default (\code{waiver()}), the name of the
scale object or the name specified in \code{labs()} is used for the title. Note
that the colour cube guide can take 3 titles: one for each axis.}
\usage{
guide_colourcube(
  title = waiver(),
  title.position = NULL,
  title.theme = NULL,
  title.hjust = NULL,
  title.vjust = NULL,
  label = TRUE,
  label.theme = NULL,
  label.hjust = NULL,
  label.vjust = NULL,
  cubewidth = NULL,
  cubeheight = NULL,
  nbin = 20,
  rotate = c(1, 2, 3, 4),
  frame.colour = "black",
  frame.linewidth = 0.5,
  frame.linetype = 1,
  ticks = TRUE,
  ticks.colour = "black",
  ticks.linewidth = 0.5,
  default.unit = "line",
  order = 0,
  available_aes = c("colour", "color", "fill"),
  ...
)
}
\arguments{
\item{title}{A character string or expression indicating a title of guide.
If \code{NULL}, the title is not shown. By default
(\code{\link[ggplot2:waiver]{waiver()}}), the name of the scale object or the name
specified in \code{\link[ggplot2:labs]{labs()}} is used for the title.}

\item{title.position}{A character string indicating the position of a
title. One of "top" (default for a vertical guide), "bottom", "left"
(default for a horizontal guide), or "right."}

\item{title.theme}{A theme object for rendering the title text. Usually the
object of \code{\link[ggplot2:element]{element_text()}} is expected. By default, the theme is
specified by \code{legend.title} in \code{\link[ggplot2:theme]{theme()}} or theme.}

\item{title.hjust}{A number specifying horizontal justification of the
title text.}

\item{title.vjust}{A number specifying vertical justification of the title
text.}

\item{label}{logical. If \code{TRUE} then the labels are drawn. If
\code{FALSE} then the labels are invisible.}

\item{label.theme}{A theme object for rendering the label text. Usually the
object of \code{\link[ggplot2:element]{element_text()}} is expected. By default, the theme is
specified by \code{legend.text} in \code{\link[ggplot2:theme]{theme()}}.}

\item{label.hjust}{A numeric specifying horizontal justification of the
label text.}

\item{label.vjust}{A numeric specifying vertical justification of the label
text.}

\item{cubewidth, cubeheight}{A \code{numeric(1)} or \code{grid::unit()} object
specifying the width/height of the colourcube. Default value is the
\code{legend.key.width/height} or \code{legend.key.size} in the theme. The colour
cube guide takes the lesser of the width/height for the size.}

\item{nbin}{A numeric specifying the number of bins for drawing the
colourbar. A smoother colourbar results from a larger value.}

\item{rotate}{NOT IMPLEMENTED YET.}

\item{frame.colour}{A string specifying the colour of the frame
drawn around the bar. If \code{NULL} (the default), no frame is drawn.}

\item{frame.linewidth}{A numeric specifying the width of the frame
drawn around the bar.}

\item{frame.linetype}{A numeric specifying the linetype of the frame
drawn around the bar.}

\item{ticks}{A logical specifying if tick marks on the colourbar should be
visible.}

\item{ticks.colour}{A string specifying the colour of the tick marks.}

\item{ticks.linewidth}{A numeric specifying the width of the tick marks.}

\item{default.unit}{A character string indicating \code{\link[grid:unit]{grid::unit()}}
for \code{barwidth} and \code{barheight}.}

\item{order}{positive integer less than 99 that specifies the order of
this guide among multiple guides. This controls the order in which
multiple guides are displayed, not the contents of the guide itself.
If 0 (default), the order is determined by a secret algorithm.}

\item{available_aes}{A vector of character strings listing the aesthetics
for which a colourbar can be drawn.}

\item{...}{ignored.}
}
\value{
A \code{guide_colourcube} S3 object.
}
\description{
Chromatic colour cube guide
}
\details{
The colour cube guide is a specialised guide for chromatic scales. It maps
three channels of a chromatic scale along three dimensions and renders an
isometric cube displaying the colours.
}
\examples{
NULL
}
